sdk 解决循环依赖和类型不匹配的问题
