+++ src/scene_server/auth_server/sdk/client/type.go
@@ -15,42 +15,7 @@ package client
 import (
 	"errors"
 	"sync"
-
-	"configcenter/src/apimachinery/util"
-	"github.com/prometheus/client_golang/prometheus"
 )
-
-type IamConfig struct {
-	// blueking's auth center addresses
-	Address []string
-	// app code is used for authorize used.
-	AppCode string
-	// app secret is used for authorized
-	AppSecret string
-	// the system id which used in auth center.
-	SystemID string
-	// http TLS config
-	TLS util.TLSClientConfig
-}
-
-func (a IamConfig) Validate() error {
-	if len(a.Address) == 0 {
-		return errors.New("no iam address")
-	}
-
-	if len(a.AppCode) == 0 {
-		return errors.New("no iam app code")
-	}
-
-	if len(a.AppSecret) == 0 {
-		return errors.New("no iam app secret")
-	}
-	return nil
-}
-
-type Options struct {
-	Metric prometheus.Registerer
-}
 
 type acDiscovery struct {
 	// auth's servers address, must prefixed with http:// or https://
